removed html rendering and moved firebase app init to routes/utils.js
