+++ index.js
@@ -4,8 +4,6 @@ const fs = require("fs");
 const express = require("express");
 const cookieParser = require("cookie-parser");
 const csrf = require("csurf");
-const auth = require('firebase/auth');
-const admin = require('firebase-admin');
 const bodyParser = require("body-parser");
 const lesson = require("./routes/lesson");
 const reflection = require("./routes/reflection");
@@ -20,13 +18,8 @@ const db = require("./model/db");
 const verseInfo = require("./routes/verseInfo");
 const tafsir = require("./routes/tafsir");
 const authentication = require("./routes/auth");
+const public = require("./routes/public")
 const { checkAuth } = require("./routes/utils");
-
-const serviceAccount = require("./serviceAccountKey.json");
-
-admin.initializeApp({
-    credential: admin.credential.cert(serviceAccount)
-});
 
 const csrfMiddleware = csrf({ cookie: true });
 
@@ -38,13 +31,14 @@ app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 app.use(cookieParser());
 // app.use(csrfMiddleware);
-
 app.engine("html", require("ejs").renderFile);
+
 
 app.use(async (req, res, next) => {
     console.log(`\nEndpoint Hit: ${req.method} ${req.originalUrl}\n`);
     next();
 });
+
 
 app.use("/api", lesson);
 app.use("/api", reflection);
@@ -55,6 +49,7 @@ app.use("/api", word);
 app.use("/api", verseInfo);
 app.use("/api", tafsir);
 app.use("/api", checkAuth, authentication);
+app.use("/auth", public);
 
 // app.all("*", (req, res, next) => {
 //     res.cookie("XSRF-TOKEN", req.csrfToken());
@@ -92,5 +87,3 @@ if (process.env.NODE_ENV == "production") {
         await setup.seedDatabase(db, true);
     });
 }
-
-
